chore #43 : aichatting, mentorchatting -> chatting-XX 모듈로 이름변경
