+++ .idea/gradle.xml
@@ -33,11 +33,11 @@
             <option value="$PROJECT_DIR$/core/model" />
             <option value="$PROJECT_DIR$/core/network" />
             <option value="$PROJECT_DIR$/feature" />
-            <option value="$PROJECT_DIR$/feature/aichatting" />
             <option value="$PROJECT_DIR$/feature/auth" />
+            <option value="$PROJECT_DIR$/feature/chatting-ai" />
+            <option value="$PROJECT_DIR$/feature/chatting-mentoring" />
             <option value="$PROJECT_DIR$/feature/consulting" />
             <option value="$PROJECT_DIR$/feature/home" />
-            <option value="$PROJECT_DIR$/feature/mentorchatting" />
             <option value="$PROJECT_DIR$/feature/mentoring" />
             <option value="$PROJECT_DIR$/feature/mentoring-mentee" />
             <option value="$PROJECT_DIR$/feature/mentoring-mentor" />
