+++ app/build.gradle.kts
@@ -56,9 +56,9 @@ dependencies {
     implementation(project(":feature:home"))
     implementation(project(":feature:splash"))
     implementation(project(":feature:consulting"))
-    implementation(project(":feature:aichatting"))
+    implementation(project(":feature:chatting-ai"))
     implementation(project(":feature:mentoring"))
-    implementation(project(":feature:mentorchatting"))
+    implementation(project(":feature:chatting-mentoring"))
     implementation(project(":feature:mentoring-mentee"))
     implementation(project(":feature:mentoring-mentor"))
     implementation(project(":feature:storage"))
