+++ feature/chatting-ai/build.gradle.kts
@@ -1,0 +1,38 @@
+plugins {
+    id("plugin.android.feature")
+    id("plugin.android.hilt")
+}
+
+android {
+    namespace = "com.tgyuu.feature.chatting.ai"
+
+    defaultConfig {
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
+    }
+
+    buildTypes {
+        release {
+            isMinifyEnabled = false
+            proguardFiles(
+                getDefaultProguardFile("proguard-android-optimize.txt"),
+                "proguard-rules.pro",
+            )
+        }
+    }
+}
+
+dependencies {
+    implementation(project(":core:domain"))
+    implementation(project(":core:model"))
+    implementation(project(":core:designsystem"))
+    implementation(project(":core:common"))
+
+    implementation(libs.lottie.compose)
+    implementation(libs.bundles.androidx)
+    implementation(libs.material)
+
+    testImplementation(libs.junit)
+    androidTestImplementation(libs.androidx.test.junit)
+    androidTestImplementation(libs.androidx.test.espresso)
+}
