+++ feature/chatting-ai/src/main/AndroidManifest.xml
