+++ feature/chatting-ai/src/main/res/drawable/ic_ai_chat_background.xml
