+++ feature/chatting-ai/src/test/java/com/tgyuu/chatting/ExampleUnitTest.kt
@@ -1,0 +1,16 @@
+package com.tgyuu.chatting
+
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
