+++ feature/chatting-mentoring/src/main/AndroidManifest.xml
