+++ feature/chatting-mentoring/src/main/res/drawable/ic_chatting_background.xml
