+++ feature/chatting-mentoring/src/main/res/drawable/ic_mentor_character.xml
