+++ feature/chatting-mentoring/src/main/res/drawable/ic_stars.xml
