+++ feature/chatting-mentoring/src/main/res/drawable/ic_yellow_stars.xml
