--- feature/mentorchatting/src/test/java/com/tgyuu/feature/mentorchatting/ExampleUnitTest.kt
@@ -1,16 +1,0 @@
-package com.tgyuu.feature.mentorchatting
-
-import junit.framework.TestCase.assertEquals
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
