+++ settings.gradle.kts
@@ -28,12 +28,12 @@ include(":feature:auth")
 include(":feature:home")
 include(":feature:splash")
 include(":feature:consulting")
-include(":feature:aichatting")
+include(":feature:chatting-ai")
 include(":feature:profile")
 include(":feature:mentoring")
 include(":feature:storage")
 include(":feature:shop")
 include(":core:database")
 include(":feature:mentoring-mentee")
-include(":feature:mentorchatting")
+include(":feature:chatting-mentoring")
 include(":feature:mentoring-mentor")
